+++ sql/queries/feed_follows.sql
@@ -32,10 +32,12 @@ INNER JOIN users
 WHERE feed_follows.user_id = $1;
 
 -- name: DeleteFeedFollow :exec
-DELETE FROM feed_follows 
-WHERE feed_id = (
-    SELECT feeds.id
-    FROM feeds
-    WHERE feeds.url = $1
-)
-AND feed_follows.user_id = $2;
+DELETE FROM feed_follows WHERE feed_id = $1 AND user_id = $2;
+--
+-- DELETE FROM feed_follows 
+-- WHERE feed_id = (
+--     SELECT feeds.id
+--     FROM feeds
+--     WHERE feeds.url = $1
+-- )
+-- AND feed_follows.user_id = $2;
